Fixed compiler error. Added test for finding examples.

+++ trunk/autosparql/src/main/java/org/dllearner/autosparql/client/model/Endpoint.java
@@ -21,7 +21,7 @@ public class Endpoint extends BaseModel{
 	}
 	
 	public int getID(){
-		return get("id");
+		return (Integer)get("id");
 	}
 
 }
